requirements.txt
# Core Dependencies
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.24.0

# Data & Analysis
yfinance>=0.2.28
textblob>=0.17.1
requests>=2.31.0

# Deep Learning
torch>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
